LPS-74523 Moves test to asset-test module
