+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/model/test/AssetVocabularyDisplayTest.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.portlet.asset.model;
+package com.liferay.asset.model.test;
 
-import com.liferay.asset.kernel.model.AssetCategoryDisplay;
+import com.liferay.asset.kernel.model.AssetVocabularyDisplay;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -22,25 +22,26 @@ import org.junit.Test;
 /**
  * @author Igor Spasic
  */
-public class AssetCategoryDisplayTest {
+public class AssetVocabularyDisplayTest {
 
 	@Test
 	public void testGetPage() {
-		AssetCategoryDisplay assetCategoryDisplay = new AssetCategoryDisplay();
+		AssetVocabularyDisplay assetVocabularyDisplay =
+			new AssetVocabularyDisplay();
 
-		assetCategoryDisplay.setStart(0);
-		assetCategoryDisplay.setEnd(20);
+		assetVocabularyDisplay.setStart(0);
+		assetVocabularyDisplay.setEnd(20);
 
-		Assert.assertEquals(1, assetCategoryDisplay.getPage());
+		Assert.assertEquals(1, assetVocabularyDisplay.getPage());
 
-		assetCategoryDisplay.setStart(20);
-		assetCategoryDisplay.setEnd(40);
+		assetVocabularyDisplay.setStart(20);
+		assetVocabularyDisplay.setEnd(40);
 
-		Assert.assertEquals(2, assetCategoryDisplay.getPage());
+		Assert.assertEquals(2, assetVocabularyDisplay.getPage());
 
-		assetCategoryDisplay.setEnd(0);
+		assetVocabularyDisplay.setEnd(0);
 
-		Assert.assertEquals(0, assetCategoryDisplay.getPage());
+		Assert.assertEquals(0, assetVocabularyDisplay.getPage());
 	}
 
 }
